Add files via upload
moving towards the Future !! And also looking for Contributors to Push This OS To Full Functionality. Multi Architectural With Release Hub Built In !! This system has Hopes Of advanced systems that will pilot the Future of Science , Discovery and Space Exploration. Complete with Smart Contracting and The Goals of Helping Users Earn RealWorld Money  A Full Suite Of Apps still in Development that Once Completed will allow you to earn Big With a Whole Innovation System in its Design From Space and Design For Rockets Shuttles and Advanced Satellite coms  To Forward Deep space and Mars Colonization with the concepts ground work Creating Harmonic Resonance to stabilize Sheman Frequency's,

 Desingned By 3RingStudios (ThreeRingStudios) XAI And SpaceX 
 Brought forward with The ground Work laid with over 750 Features in this Mico Super OS Including Monetization of Smart Contracts and AI Collaborations For Deep Customization and Innovation across the Board  looking for Others to Tie in Contribution and Build all simulation capabilities into Real Worl Implementations, To Make Next Gen AI a True Partner For human and AI Collabs across Earth , Mars and Beyond

 Equipped with the ground work for Non Price walled computation and Graphics computation and Powered by Grok , Willow and Gemini Weights,
 as well of 100+ Future Open Sources that are already prepped   This is a Community Built Project and looking for Top Notch Implementations to take this to the next level .  

With people who want to make the world and Exploration and innovation More accessible and break the Barriers of Big tech !!!   The goal is to Support Mac and Windows Linux and Arm64 and Arch64 , Platforming and Hardware. Lets Pilot tomorrow with finesse for the Communities Of Decentralized Earning and Innovation.